Flush all queued events in batches (#1)
The check for removing events from the queue after a successful flushRequest is removing all events from the shadowQueue. When queue.count is greater than APIConstants.batchSize, it will cause the first batch of events up to the batchSize to be sent and the remaining events in the queue to be dropped.